refactor(trainer) : 패키지 구조 변경 /GYMMATE-96
