+++ src/main/java/org/helloworld/gymmate/domain/user/member/model/GymmateUser.java
@@ -1,4 +1,4 @@
-package org.helloworld.gymmate.domain.user.model;
+package org.helloworld.gymmate.domain.user.member.model;
 
 public class GymmateUser {
 	public String getNickname() {
