--- src/main/java/org/helloworld/gymmate/domain/user/service/GymmateUserDetailsService.java
@@ -1,4 +1,0 @@
-package org.helloworld.gymmate.domain.user.service;
-
-public class GymmateUserDetailsService {
-}
