+++ src/main/java/org/helloworld/gymmate/domain/user/trainer/controller/GymMateTrainerController.java
@@ -1,6 +1,6 @@
-package org.helloworld.gymmate.domain.user.controller;
+package org.helloworld.gymmate.domain.user.trainer.controller;
 
-import org.helloworld.gymmate.domain.user.service.GymMateTrainerService;
+import org.helloworld.gymmate.domain.user.trainer.service.GymMateTrainerService;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
